added sales tab
